pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
[project]
name = "video-pipelines"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckdb>=1.1.3",
    "ffmpeg-python>=0.2.0",
    "jinja2>=3.1.6",
    "pandas>=2.2.3",
    "pydantic>=2.11.7",
    "typer>=0.16.0",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = ["ipykernel>=6.30.1", "pip>=25.2", "python-dotenv>=1.1.1"]
lint = ["mypy>=1.17.1", "ruff>=0.12.7", "types-pyyaml>=6.0.12.20250516"]
format = ["ruff>=0.12.7"]

[tool.mypy]
files = "src"
